+++ bidsme/Modules/MRI/jsonNIFTI.py
@@ -24,7 +24,6 @@
 
 from ..common import retrieveFormDict
 from .MRI import MRI
-from bidsMeta import MetaField
 
 import os
 import logging
@@ -40,9 +39,8 @@ class jsonNIFTI(MRI):
 
     __slots__ = ["_HEADER_CACHE", "_FILE_CACHE",
                  "_header_file",
-                 "manufacturer",
                  ]
-    __specialFields = {"acq_time"}
+    __specialFields = {}
 
     def __init__(self, rec_path=""):
         super().__init__()
@@ -50,96 +48,9 @@ class jsonNIFTI(MRI):
         self._HEADER_CACHE = None
         self._FILE_CACHE = ""
         self._header_file = ""
-        self.manufacturer = ""
 
         if rec_path:
             self.setRecPath(rec_path)
-
-        #####################
-        # Common recomended #
-        #####################
-        recommend = self.metaFields_rec["__common__"]
-        recommend["Manufacturer"] = MetaField("Manufacturer")
-        recommend["ManufacturersModelName"] =\
-            MetaField("ManufacturerModelName")
-        recommend["DeviceSerialNumber"] = MetaField("DeviceSerialNumber")
-        recommend["StationName"] = MetaField("StationName")
-        recommend["SoftwareVersions"] = MetaField("SoftwareVersions")
-        recommend["MagneticFieldStrength"] =\
-            MetaField("MagneticFieldStrength", 1.)
-        recommend["ReceiveCoilName"] = MetaField("ReceiveCoilName")
-        recommend["ReceiveCoilActiveElements"] =\
-            MetaField("ReceiveCoilActiveElements")
-        # recommend["GradientSetType"]
-        recommend["MRTransmitCoilSequence"] =\
-            MetaField("MRTransmitCoilSequence")
-        # recommend["MatrixCoilMode"]
-        # recommend["CoilCombinationMethod"]
-        # recommend["PulseSequenceType"]
-        recommend["ScanningSequence"] = MetaField("ScanningSequence")
-        recommend["SequenceVariant"] = MetaField("SequenceVariant")
-        recommend["ScanOptions"] = MetaField("ScanOptions")
-        recommend["SequenceName"] = MetaField("SequenceName")
-        # recommend["NumberShots"]
-        # recommend["PulseSequenceDetails"]
-        # recommend["NonlinearGradientCorrection"]
-        # recommend["ParallelReductionFactorInPlane"]
-        # recommend["ParallelAcquisitionTechnique"]
-        recommend["PartialFourier"] = MetaField("PartialFourier")
-        recommend["PartialFourierDirection"] =\
-            MetaField("PartialFourierDirection")
-        # recommend["PhaseEncodingDirection"]
-        # recommend["EffectiveEchoSpacing"]
-        # recommend["TotalReadoutTime"]
-        recommend["EchoTime"] = MetaField("EchoTime", 1e-3)
-        recommend["InversionTime"] = MetaField("InversionTime", 1e-3)
-        # recommend["SliceTiming"]
-        # recommend["SliceEncodingDirection"]
-        # recommend["DwellTime"] = MetaField("Private_0019_1018", 1e-6)
-        recommend["FlipAngle"] = MetaField("FlipAngle", 1.)
-        # recommend["MultibandAccelerationFactor"]
-        # recommend["NegativeContrast"]
-        # recommend["MultibandAccelerationFactor"]
-        # recommend["AnatomicalLandmarkCoordinates"]
-        recommend["InstitutionName"] = MetaField("InstitutionName")
-        recommend["InstitutionAddress"] = MetaField("InstitutionAddress")
-        recommend["InstitutionalDepartmentName"] =\
-            MetaField("InstitutionalDepartmentName")
-
-        #####################
-        # sMRI metafields   #
-        #####################
-        # optional = self.metaFields_opt["anat"]
-        # optional["ContrastBolusIngredient"]
-
-        #####################
-        # fMRI metafields   #
-        #####################
-        required = self.metaFields_req["func"]
-        required["RepetitionTime"] = MetaField("RepetitionTime", 1e-3)
-        required["TaskName"] = MetaField("<<bids:task>>")
-
-        recommend = self.metaFields_rec["func"]
-        # recommend["NumberOfVolumesDiscardedByScanner"]
-        # recommend["NumberOfVolumesDiscardedByUser"]
-        # recommend["DelayTime"]
-        # recommend["AcquisitionDuration"]
-        # recommend["DelayAfterTrigger"]
-        # recommend["Instructions"]
-        # recommend["TaskDescription"]
-        # recommend["CogAtlasID"]
-        # recommend["CogPOID"]
-
-        #####################
-        # fMap metafields   #
-        #####################
-        required = self.metaFields_req["fmap"]
-        # required["IntendedFor"]
-
-        recommend = self.metaFields_rec["fmap"]
-        # recommend["EchoTime1"]
-        # recommend["EchoTime2"]
-        # recommend["Units"]
 
     ################################
     # reimplementation of virtuals #
@@ -191,10 +102,9 @@ class jsonNIFTI(MRI):
             self._FILE_CACHE = path
             self._HEADER_CACHE = dicomdict
             self._header_file = header
-            self.manufacturer = self.getField("Manufacturer", "")
 
     def acqTime(self) -> datetime:
-        return self.getAttribute("acq_time")
+        return None
 
     def dump(self):
         if self._HEADER_CACHE is not None:
@@ -221,18 +131,10 @@ class jsonNIFTI(MRI):
         return res
 
     def recNo(self):
-        return self.getField("SeriesNumber", 0)
+        return self.index
 
     def recId(self):
-        seriesdescr = self.getField("SeriesDescription")
-        if seriesdescr is None:
-            seriesdescr = self.getField("ProtocolName")
-        if seriesdescr is None:
-            logger.warning("{}: Unable to get recording Id for file {}"
-                           .format(self.formatIdentity(),
-                                   self.currentFile()))
-            seriesdescr = "unknown"
-        return seriesdescr.strip()
+        return os.path.splitext(self.currentFile(True))[0]
 
     def isCompleteRecording(self):
         return True
@@ -251,7 +153,7 @@ class jsonNIFTI(MRI):
         shutil.copy2(self._header_file, destination)
 
     def _getSubId(self) -> str:
-        return str(self.getField("PatientID"))
+        return ""
 
     def _getSesId(self) -> str:
         return ""
@@ -260,73 +162,4 @@ class jsonNIFTI(MRI):
     # Additional fonctions #
     ########################
     def _adaptMetaField(self, name):
-        if name == "acq_time":
-            if "AcquisitionDateTime" in self._HEADER_CACHE:
-                dt_stamp = self._HEADER_CACHE["AcquisitionDateTime"]
-                return self.__transform(dt_stamp, "DT")
-
-            acq = datetime.min
-
-            if "AcquisitionDate" in self._HEADER_CACHE:
-                date_stamp = self.__transform(
-                        self._HEADER_CACHE["AcquisitionDate"],
-                        "DA"
-                        )
-            else:
-                logger.warning("{}: Acquisition Date not defined"
-                               .format(self.recIdentity()))
-                return acq
-
-            if "AcquisitionTime" in self._HEADER_CACHE:
-                time_stamp = self.__transform(
-                        self._HEADER_CACHE["AcquisitionTime"],
-                        "TM")
-            else:
-                logger.warning("{}: Acquisition Time not defined"
-                               .format(self.recIdentity()))
-                return acq
-            acq = datetime.combine(date_stamp, time_stamp)
-            return acq
-
         return None
-
-    def __transform(self, value, VR: str):
-        # Date and time
-        # converted to corresponding datetime subclass
-        if VR == "TM":
-            if "." in value:
-                dt = datetime.strptime(value, "%H:%M:%S.%f").time()
-            else:
-                dt = datetime.strptime(value, "%H:%M:%S").time()
-            return dt
-        if VR == "DA":
-            dt = datetime.strptime(value, "%Y-%m-%d").date()
-            return dt
-        if VR == "DT":
-            value = value.strip()
-            date_string = "%Y-%m-%d"
-            time_string = "%H:%M:%S"
-            sep_string = "T"
-            ms_string = ""
-            uts_string = ""
-            if "." in value:
-                ms_string += ".%f"
-            if "+" in value or "-" in value:
-                uts_string += "%z"
-            if len(value) == 8 or \
-                    (len(value) == 13 and uts_string != ""):
-                logger.warning("{}: Format is DT, but string looks like DA"
-                               .format(value))
-                t = datetime.strptime(value, date_string + uts_string)
-            elif len(value) == 6 or \
-                    (len(value) == 13 and ms_string != ""):
-                logger.warning("{}: Format is DT, but string looks like TM"
-                               .format(value))
-                t = datetime.strptime(value, time_string + ms_string)
-            else:
-                t = datetime.strptime(value, date_string + sep_string
-                                      + time_string + ms_string + uts_string)
-            if t.tzinfo is not None:
-                t += t.tzinfo.utcoffset(t)
-                t = t.replace(tzinfo=None)
-            return t